api/requirements.txt
openai==0.27.0
spotipy==2.20.0
Flask==3.1.0
python-dotenv==1.0.0

